@@ -16,91 +17,106 @@
   "dependencies": {
     "@johmun/vue-tags-input": "^2.1.0",
     "@telnyx/webrtc": "^2.5.6",
+    "@vue/compat": "^3.3.4",
     "axios": ">=0.21.2",
     "bootstrap": "^5.1.0",
     "bootstrap-vue": "^2.21.2",
     "cbor-js": "^0.1.0",
-    "cbor-x": "^1.0.0",
+    "cbor-x": "^1.5.3",
     "gsap": "^3.8.0",
     "papaparse": "^5.3.1",
     "socket.io-client": "^4.1.3",
-    "vue": "^2.5.2",
+    "vue": "^3.3.4",
     "vue-axios": "^3.3.6",
-    "vue-chat-scroll": "^1.4.0",
     "vue-cookie": "^1.1.4",
     "vue-input-tag": "^2.0.7",
     "vue-moment": "^4.1.0",
     "vue-router": "^3.0.1",
-    "vue-select": "^3.16.0",
+    "vue-select": "^4.0.0-beta.6",
     "vue-sweetalert2": "^5.0.2",
     "vue2-form-loading": "^0.1.0",
-    "vuelidate": "^0.7.6",
-    "vuex": "^3.6.2",
+    "vuelidate": "^0.7.7",
+    "vuex": "^4.1.0",
     "zlib": "^1.0.5"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.2",
-    "babel-core": "^6.22.1",
-    "babel-eslint": "^8.2.1",
+    "@babel/core": "^7.0.0",
+    "@babel/eslint-parser": "^7.22.5",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-function-sent": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/plugin-syntax-jsx": "^7.0.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.22.5",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "@soda/friendly-errors-webpack-plugin": "^1.8.1",
+    "@vue/babel-helper-vue-jsx-merge-props": "^1.4.0",
+    "@vue/babel-plugin-transform-vue-jsx": "^1.4.0",
+    "@vue/babel-preset-jsx": "^1.4.0",
+    "@vue/compiler-sfc": "^3.3.4",
+    "@vue/vue3-jest": "^28.1.0",
+    "autoprefixer": "^10.4.14",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^10.1.0",
     "babel-helper-vue-jsx-merge-props": "^2.0.3",
-    "babel-jest": "^21.0.2",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-dynamic-import-node": "^1.2.0",
-    "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-plugin-transform-vue-jsx": "^3.5.0",
-    "babel-preset-env": "^1.3.2",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "chalk": "^2.0.1",
+    "babel-jest": "^28.1.3",
+    "babel-loader": "^9.1.2",
+    "babel-plugin-dynamic-import-node": "^2.3.3",
+    "chalk": "^4.1.2",
     "chromedriver": "^2.27.2",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-spawn": "^5.0.1",
-    "css-loader": "^0.28.0",
-    "eslint": "^4.15.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.2.0",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "eslint-plugin-vue": "^4.0.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^1.1.4",
-    "friendly-errors-webpack-plugin": "^1.6.1",
-    "html-webpack-plugin": "^2.30.1",
-    "jest": "^22.0.4",
-    "jest-serializer-vue": "^0.3.0",
-    "nightwatch": "^0.9.12",
-    "node-notifier": "^5.1.2",
-    "optimize-css-assets-webpack-plugin": "^3.2.0",
-    "ora": "^1.2.0",
-    "portfinder": "^1.0.13",
-    "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.0.8",
-    "postcss-url": "^7.2.1",
+    "copy-webpack-plugin": "^11.0.0",
+    "cross-spawn": "^7.0.3",
+    "css-loader": "^6.8.1",
+    "css-minimizer-webpack-plugin": "^5.0.1",
+    "eslint": "^8.43.0",
+    "eslint-config-promise": "^1.2.6",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^16.0.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-standard": "^5.0.0",
+    "eslint-plugin-vue": "^9.15.1",
+    "eslint-webpack-plugin": "^4.0.1",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.3",
+    "jest": "^28.1.3",
+    "jest-serializer-vue": "^3.1.0",
+    "mini-css-extract-plugin": "^2.7.6",
+    "nightwatch": "^2.0.10",
+    "node-notifier": "^10.0.1",
+    "nodemon": "^2.0.22",
+    "ora": "^5.4.1",
+    "portfinder": "^1.0.32",
+    "postcss-import": "^15.1.0",
+    "postcss-loader": "^7.3.3",
+    "postcss-url": "^10.0.0",
     "pulltorefreshjs": "^0.1.22",
-    "rimraf": "^2.6.0",
     "selenium-server": "^3.0.1",
-    "semver": "^5.3.0",
+    "semver": "^7.5.3",
     "shelljs": "^0.8.5",
-    "twilio-client": "^1.14.0",
-    "uglifyjs-webpack-plugin": "^1.1.1",
-    "url-loader": "^0.5.8",
-    "vue-cli-plugin-compression": "~1.1.5",
-    "vue-jest": "^1.0.2",
-    "vue-loader": "^13.3.0",
-    "vue-style-loader": "^3.0.1",
-    "vue-template-compiler": "^2.5.2",
-    "webpack": "^3.6.0",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-server": "^2.9.1",
-    "webpack-merge": "^4.1.0"
+    "style-loader": "^3.3.3",
+    "twilio-client": "^1.15.0",
+    "url-loader": "^4.1.1",
+    "util": "^0.12.5",
+    "vue-cli-plugin-compression": "~2.0.0",
+    "vue-loader": "^17.2.2",
+    "vue-style-loader": "^4.1.3",
+    "webpack": "^5.88.0",
+    "webpack-bundle-analyzer": "^4.9.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^5.9.0"
   },
   "engines": {
-    "node": ">= 6.0.0",
+    "node": ">= 12.13.0",
     "npm": ">= 3.0.0"
   },
   "browserslist": [
