@@ -1,5 +1,5 @@
 'use strict'
-const merge = require('webpack-merge')
+const { merge } = require('webpack-merge')
 const prodEnv = require('./prod.env')
 
 module.exports = merge(prodEnv, {
