@@ -1,5 +1,5 @@
 'use strict'
-const merge = require('webpack-merge')
+const { merge } = require('webpack-merge')
 const devEnv = require('./dev.env')
 
 module.exports = merge(devEnv, {
